requirements.txt
flipside==2.0.8
pandas==2.2.2
requests==2.29.0
python-dotenv==1.0.1
pyyaml==6.0.1
dune-client==1.7.3
backoff==2.2.1